+++ example/example/lib/stories.dart
@@ -1,7 +1,7 @@
 import 'package:example/stories/zwap_buttons_story.dart';
 import 'package:example/stories/zwap_checkbox_story.dart';
 import 'package:example/stories/zwap_input_story.dart';
-import 'package:example/stories/zwap_percent.dart';
+import 'package:example/stories/zwap_percent_story.dart';
 import 'package:example/stories/zwap_select_story.dart';
 import 'package:example/stories/zwap_weekly_calendar_pickert.dart';
 import 'package:flutter/material.dart';